replace Integer with BigInteger

+++ src/node/DataOutNode.java
@@ -3,6 +3,7 @@ package node;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.PrintStream;
+import java.math.BigInteger;
 import java.util.List;
 
 public class DataOutNode implements Node {
@@ -22,7 +23,7 @@ public class DataOutNode implements Node {
 	}
 
 	@Override
-	public Integer evaluate(List<Integer> inputs) {
+	public BigInteger evaluate(List<BigInteger> inputs) {
 		if (inputs == null || inputs.size() != 1)
 			throw new IllegalArgumentException();
 		out.println(inputs.getFirst()+",");
